refactor: index.ts로 import문 간단하게 작성
