+++ backend/was/src/auth/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './jwt-payload.dto';
+export * from './oauth-token.dto';
+export * from './profile.dto';
