+++ backend/was/src/auth/dto/kakao/index.ts
@@ -1,0 +1,2 @@
+export * from './kakao-token.dto';
+export * from './kakao-access-token-info.dto';
