+++ backend/was/src/auth/service/kakao.auth.service.ts
@@ -11,13 +11,10 @@ import { Cache } from 'cache-manager';
 import { CONTENT_TYPE, METHODS, OAUTH_URL } from 'src/common/constants/apis';
 import { ERR_MSG } from 'src/common/constants/errors';
 import { PROVIDER_ID, PROVIDER_NAME } from 'src/common/constants/etc';
-import { Member } from 'src/members/entities/member.entity';
+import { Member } from 'src/members/entities';
 import { MembersService } from 'src/members/members.service';
-import { JwtPayloadDto } from '../dto/jwt-payload.dto';
-import { KakaoAccessTokenInfoDto } from '../dto/kakao/kakao-access-token-info.dto';
-import { KakaoTokenDto } from '../dto/kakao/kakao-token.dto';
-import { OAuthTokenDto } from '../dto/oauth-token.dto';
-import { ProfileDto } from '../dto/profile.dto';
+import { JwtPayloadDto, OAuthTokenDto, ProfileDto } from '../dto';
+import { KakaoAccessTokenInfoDto, KakaoTokenDto } from '../dto/kakao';
 import { CacheKey } from '../interface/cache-key';
 import { makeRefreshTokenForm, makeRequestTokenForm } from '../util/kakao';
 import { AuthService } from './auth.service';
