+++ backend/was/src/chat/dto/index.ts
@@ -1,0 +1,4 @@
+export * from './chatting-message-response.dto';
+export * from './chatting-room-response.dto';
+export * from './create-chatting-message.dto';
+export * from './update-chatting-room.dto';
