+++ backend/was/src/chat/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './chatting-message.entity';
+export * from './chatting-room.entity';
