+++ backend/was/src/members/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-member.dto';
+export * from './update-member.dto';
