+++ backend/was/src/members/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './member.entity';
