+++ backend/was/src/members/members.module.ts
@@ -1,12 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Member } from './entities/member.entity';
-// import { MembersController } from './members.controller';
+import { Member } from './entities';
 import { MembersService } from './members.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Member])],
-  // controllers: [MembersController],
   providers: [MembersService],
 })
 export class MembersModule {}
