+++ backend/was/src/tarot/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './create-tarot-result.dto';
+export * from './tarot-card.dto';
+export * from './tarot-result.dto';
