+++ backend/was/src/tarot/dto/tarot-card.dto.ts
@@ -1,14 +1,14 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsUrl } from 'class-validator';
 import { BUCKET_URL } from 'src/common/constants/etc';
-import { TarotCard } from '../entities/tarot-card.entity';
+import { TarotCard } from '../entities';
 
-export class TarotCardResponseDto {
+export class TarotCardDto {
   @IsUrl()
   @ApiProperty({ description: '타로 카드 이미지 URL', required: true })
   readonly cardUrl: string;
 
-  static fromEntity(entity: TarotCard): TarotCardResponseDto {
+  static fromEntity(entity: TarotCard): TarotCardDto {
     return { cardUrl: `${BUCKET_URL}/basic/${entity.cardNo}${entity.ext}` };
   }
 }
