+++ backend/was/src/tarot/dto/tarot-result.dto.ts
@@ -1,8 +1,8 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsString, IsUrl } from 'class-validator';
-import { TarotResult } from '../entities/tarot-result.entity';
+import { TarotResult } from '../entities';
 
-export class TarotResultResponseDto {
+export class TarotResultDto {
   @IsUrl()
   @ApiProperty({ description: '타로 카드 이미지 URL', required: true })
   readonly cardUrl: string;
@@ -11,7 +11,7 @@ export class TarotResultResponseDto {
   @ApiProperty({ description: '타로 해설 결과', required: true })
   readonly message: string;
 
-  static fromEntity(entity: TarotResult): TarotResultResponseDto {
+  static fromEntity(entity: TarotResult): TarotResultDto {
     return { ...entity };
   }
 }
