+++ backend/was/src/tarot/entities/index.ts
@@ -1,0 +1,3 @@
+export * from './tarot-card-pack.entity';
+export * from './tarot-card.entity';
+export * from './tarot-result.entity';
