+++ backend/was/src/tarot/tarot.module.ts
@@ -1,8 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TarotCardPack } from './entities/tarot-card-pack.entity';
-import { TarotCard } from './entities/tarot-card.entity';
-import { TarotResult } from './entities/tarot-result.entity';
+import { TarotCard, TarotCardPack, TarotResult } from './entities';
 import { TarotController } from './tarot.controller';
 import { TarotService } from './tarot.service';
