+++ backend/was/src/tarot/tarot.service.ts
@@ -2,11 +2,8 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ERR_MSG } from 'src/common/constants/errors';
 import { Repository } from 'typeorm';
-import { CreateTarotResultDto } from './dto/create-tarot-result.dto';
-import { TarotCardResponseDto } from './dto/tarot-card-response.dto';
-import { TarotResultResponseDto } from './dto/tarot-result-response.dto';
-import { TarotCard } from './entities/tarot-card.entity';
-import { TarotResult } from './entities/tarot-result.entity';
+import { CreateTarotResultDto, TarotCardDto, TarotResultDto } from './dto';
+import { TarotCard, TarotResult } from './entities';
 
 @Injectable()
 export class TarotService {
@@ -33,7 +30,7 @@ export class TarotService {
   /**
    * TODO : 추후 타로 카드팩이 커스텀이 가능한 경우, 전체적인 로직 수정 필요
    */
-  async findTarotCardById(id: number): Promise<TarotCardResponseDto> {
+  async findTarotCardById(id: number): Promise<TarotCardDto> {
     const tarotCard: TarotCard | null =
       await this.tarotCardRepository.findOneBy({
         cardNo: id,
@@ -42,15 +39,15 @@ export class TarotService {
     if (!tarotCard) {
       throw new NotFoundException(ERR_MSG.TAROT_CARD_NOT_FOUND);
     }
-    return TarotCardResponseDto.fromEntity(tarotCard);
+    return TarotCardDto.fromEntity(tarotCard);
   }
 
-  async findTarotResultById(id: string): Promise<TarotResultResponseDto> {
+  async findTarotResultById(id: string): Promise<TarotResultDto> {
     const tarotResult: TarotResult | null =
       await this.tarotResultRepository.findOneBy({ id });
     if (!tarotResult) {
       throw new NotFoundException(ERR_MSG.TAROT_RESULT_NOT_FOUND);
     }
-    return TarotResultResponseDto.fromEntity(tarotResult);
+    return TarotResultDto.fromEntity(tarotResult);
   }
 }
